+++ lib/sigil.ex
@@ -9,13 +9,19 @@ defmodule Dinero.Sigil do
   ## Usage
 
       iex> import Dinero.Sigil
-      iex> ~m[100.24]usd
-      %Dinero{amount: 10024, currency: :USD}
+      iex> ~m[100.24]uah
+      %Dinero{amount: 10024, currency: :UAH}
+      iex> ~m[100]
+      %Dinero{amount: 10000, currency: :USD}
 
   """
 
   defmacro sigil_m({:<<>>, _, [amount]}, [_ | _] = currency) do
     Macro.escape(Dinero.new(to_float(amount), List.to_atom(currency)))
+  end
+
+  defmacro sigil_m({:<<>>, _, [amount]}, []) do
+    Macro.escape(Dinero.new(to_float(amount), :USD))
   end
 
   defp to_float(amount) do